# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для бронирования переговорных комнат",
        "title": "Booking Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/rooms/{roomId}/schedule/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создать расписание для переговорки (только admin, только один раз). Длительность слота 30 мин.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Создать расписание переговорки",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID комнаты",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные расписания",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ScheduleCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.ScheduleCreateResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/rooms/{roomId}/slots/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Возвращает слоты, не занятые активной бронью, для указанной переговорки на указанную дату",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Slots"
                ],
                "summary": "Список доступных слотов",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID комнаты",
                        "name": "roomId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "format": "date",
                        "description": "Дата в формате ISO 8601 (например: 2024-06-10)",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SlotListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.InternalErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ScheduleCreateRequest": {
            "type": "object",
            "properties": {
                "daysOfWeek": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "endTime": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "dto.ScheduleCreateResponse": {
            "type": "object",
            "properties": {
                "schedule": {
                    "$ref": "#/definitions/dto.ScheduleResponse"
                }
            }
        },
        "dto.ScheduleResponse": {
            "type": "object",
            "properties": {
                "daysOfWeek": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "dto.SlotListResponse": {
            "type": "object",
            "properties": {
                "slots": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.SlotResponse"
                    }
                }
            }
        },
        "dto.SlotResponse": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                }
            }
        },
...
        "server.ErrorCode": {
            "type": "string",
            "enum": [
                "INVALID_REQUEST",
                "UNAUTHORIZED",
                "NOT_FOUND",
                "ROOM_NOT_FOUND",
                "SLOT_NOT_FOUND",
                "SLOT_ALREADY_BOOKED",
                "BOOKING_NOT_FOUND",
                "FORBIDDEN",
                "SCHEDULE_EXISTS",
                "INTERNAL_ERROR"
            ],
            "x-enum-comments": {
                "CodeBookingNotFound": "в документации нет упоминания в каких кейсах используется"
            },
            "x-enum-descriptions": [
                "",
                "",
                "",
                "",
                "",
                "",
                "в документации нет упоминания в каких кейсах используется",
                "",
                "",
                ""
            ],
            "x-enum-varnames": [
                "CodeInvalidRequest",
                "CodeUnauthorized",
                "CodeNotFound",
                "CodeRoomNotFound",
                "CodeSlotNotFound",
                "CodeSlotAlreadyBooked",
                "CodeBookingNotFound",
                "CodeForbidden",
                "CodeScheduleExists",
                "CodeInternalError"
            ]
        },
        "server.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/server.ErrorCode"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/server.ErrorDetail"
                }
            }
        },
        "server.InternalErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/server.ErrorDetail"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}